Tweak run.sh for older projects

+++ 7.0/configs/run.sh
@@ -4,16 +4,26 @@ export LC_ALL="en_US.UTF-8"
 export LANG="en_US.UTF-8"
 
 cd /app
-composer install  --optimize-autoloader --ignore-platform-reqs
+if [ -f composer.json ]; then
+    composer install  --optimize-autoloader --ignore-platform-reqs
+fi
 
-cp /app/package.json /root/
-cd /root && npm install -q
-rsync -rlzuIO --ignore-errors /root/node_modules/ /app/node_modules > /dev/null 2>&1
+if [ -f package.json ]; then
+    cp /app/package.json /root/
+    cd /root && npm install -q
+    rsync -rlzuIO --ignore-errors /root/node_modules/ /app/node_modules > /dev/null 2>&1
+fi
 cd /app
 
-bundle install
-bower install --allow-root
-gulp build
+if [ -f Gemfile ]; then
+    bundle install
+fi
+if [ -f bower.json ]; then
+    bower install --allow-root
+fi
+if [ -f Gulpfile ]; then
+    gulp build
+fi
 bin/console assets:install --symlink --env=docker || app/console assets:install --symlink --env=docker
 bin/console assetic:dump --env=docker || app/console assetic:dump --env=docker
